split out dry run
